rebase to main; moving interfaces into their own files
